fix: building errors caused by update

+++ lib/shared/data/repositories/tracked_item_repository_impl.dart
@@ -174,7 +174,7 @@ class TrackedItemRepositoryImpl implements TrackedItemRepository {
           prefix: "TrackedItemRepository",
         );
 
-        totalSynced += (countAfter! - countBefore!);
+        totalSynced += (countAfter! - countBefore!).toInt();
       } catch (e) {
         Logger.error(
           "Skipping item '${item.name}' (ID: ${item.id}) or sync failed: $e",
